be_storage_t
BE data is internally stored as u8, u16, u32, u64, u128 value (decreases
amount of reinterpret casts in be<>le conversions)

+++ rpcs3/Emu/SysCalls/Modules/cellPngDec.cpp
@@ -77,7 +77,7 @@ s64 pngDecOpen(
 	stream->fd = 0;
 	stream->src = *src;
 
-	switch ((u32)src->srcSelect.ToBE())
+	switch (src->srcSelect.ToBE())
 	{
 	case se32(CELL_PNGDEC_BUFFER):
 		stream->fileSize = src->streamSize.ToLE();
@@ -145,7 +145,7 @@ s64 pngReadHeader(
 	auto buffer_32 = buffer.To<be_t<u32>>();
 	vm::var<be_t<u64>> pos, nread;
 
-	switch ((u32)stream->src.srcSelect.ToBE())
+	switch (stream->src.srcSelect.ToBE())
 	{
 	case se32(CELL_PNGDEC_BUFFER):
 		memmove(buffer.begin(), stream->src.streamPtr.get_ptr(), buffer.size());
@@ -206,7 +206,7 @@ s64 pngDecSetParameter(
 	current_outParam.outputHeight = current_info.imageHeight;
 	current_outParam.outputColorSpace = inParam->outputColorSpace;
 
-	switch ((u32)current_outParam.outputColorSpace.ToBE())
+	switch (current_outParam.outputColorSpace.ToBE())
 	{
 	case se32(CELL_PNGDEC_PALETTE):
 	case se32(CELL_PNGDEC_GRAYSCALE):
@@ -254,7 +254,7 @@ s64 pngDecodeData(
 	vm::var<unsigned char[]> png((u32)fileSize);
 	vm::var<be_t<u64>> pos, nread;
 
-	switch ((u32)stream->src.srcSelect.ToBE())
+	switch (stream->src.srcSelect.ToBE())
 	{
 	case se32(CELL_PNGDEC_BUFFER):
 		memmove(png.begin(), stream->src.streamPtr.get_ptr(), png.size());
@@ -283,7 +283,7 @@ s64 pngDecodeData(
 	const int bytesPerLine = (u32)dataCtrlParam->outputBytesPerLine;
 	uint image_size = width * height;
 
-	switch ((u32)current_outParam.outputColorSpace.ToBE())
+	switch (current_outParam.outputColorSpace.ToBE())
 	{
 	case se32(CELL_PNGDEC_RGB):
 	case se32(CELL_PNGDEC_RGBA):
